Fix navbar auth

+++ src/Controller/UserController.php
@@ -5,8 +5,12 @@ namespace App\Controller;
 use App\Entity\User;
 use App\Form\UserType;
 use App\Repository\UserRepository;
+use Exception;
+use Sensio\Bundle\FrameworkExtraBundle\Configuration\IsGranted;
+use Symfony\Component\Finder\Exception\AccessDeniedException;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\HttpKernel\Exception\AccessDeniedHttpException;
 use Symfony\Component\Routing\Annotation\Route;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\Security\Core\Encoder\UserPasswordEncoderInterface;
@@ -18,9 +22,17 @@ class UserController extends AbstractController
      */
     public function index(UserRepository $userRepository): Response
     {
-        return $this->render('user/index.html.twig', [
-            'users' => $userRepository->findAll(),
-        ]);
+        try{
+            if(!$this->getUser()) throw new AccessDeniedException();
+            $this->denyAccessUnlessGranted('ROLE_ADMIN');
+
+            return $this->render('user/index.html.twig', [
+                'users' => $userRepository->findAll(),
+            ]);
+        }
+        catch(Exception $e){
+            return $this->redirectToRoute('home');
+        }
     }
 
     /**
@@ -28,29 +40,37 @@ class UserController extends AbstractController
      */
     public function new(Request $request): Response
     {
-        $user = new User();
-        $form = $this->createForm(UserType::class, $user);
-        $form->handleRequest($request);
+        try{
+            if(!$this->getUser()) throw new AccessDeniedException();
+            $this->denyAccessUnlessGranted('ROLE_ADMIN');
 
-        if ($form->isSubmitted() && $form->isValid()) {
-            $entityManager = $this->getDoctrine()->getManager();
-            
-            // ATTENTION
-            // Encoder les mots de passe
-            // avant de les enregister dans la base
-            $hash = $encoder->encodePassword($user, $user->getPassword());
-            $user->setPassword($hash);
+            $user = new User();
+            $form = $this->createForm(UserType::class, $user);
+            $form->handleRequest($request);
 
-            $entityManager->persist($user);
-            $entityManager->flush();
+            if ($form->isSubmitted() && $form->isValid()) {
+                $entityManager = $this->getDoctrine()->getManager();
 
-            return $this->redirectToRoute('user_index');
+                // ATTENTION
+                // Encoder les mots de passe
+                // avant de les enregister dans la base
+                $hash = $encoder->encodePassword($user, $user->getPassword());
+                $user->setPassword($hash);
+
+                $entityManager->persist($user);
+                $entityManager->flush();
+
+                return $this->redirectToRoute('user_index');
+            }
+
+            return $this->render('user/new.html.twig', [
+                'user' => $user,
+                'form' => $form->createView(),
+            ]);
         }
-
-        return $this->render('user/new.html.twig', [
-            'user' => $user,
-            'form' => $form->createView(),
-        ]);
+        catch(Exception $e){
+            return $this->redirectToRoute('home');
+        }
     }
 
     /**
@@ -58,9 +78,17 @@ class UserController extends AbstractController
      */
     public function show(User $user): Response
     {
-        return $this->render('user/show.html.twig', [
-            'user' => $user,
-        ]);
+        try {
+            if (!$this->getUser()) throw new AccessDeniedException();
+            $this->denyAccessUnlessGranted('ROLE_ADMIN');
+
+            return $this->render('user/show.html.twig', [
+                'user' => $user,
+            ]);
+        }
+        catch(Exception $e){
+            return $this->redirectToRoute('home');
+        }
     }
 
     /**
@@ -68,22 +96,30 @@ class UserController extends AbstractController
      */
     public function edit(Request $request, User $user, UserPasswordEncoderInterface $encoder): Response
     {
-        $form = $this->createForm(UserType::class, $user);
-        $form->handleRequest($request);
+        try {
+            if(!$this->getUser()) throw new AccessDeniedException();
+            $this->denyAccessUnlessGranted('ROLE_ADMIN');
 
-        if ($form->isSubmitted() && $form->isValid()) {
-            $hash = $encoder->encodePassword($user, $user->getPassword());
-            $user->setPassword($hash);
+            $form = $this->createForm(UserType::class, $user);
+            $form->handleRequest($request);
 
-            $this->getDoctrine()->getManager()->flush();
+            if ($form->isSubmitted() && $form->isValid()) {
+                $hash = $encoder->encodePassword($user, $user->getPassword());
+                $user->setPassword($hash);
 
-            return $this->redirectToRoute('user_index');
+                $this->getDoctrine()->getManager()->flush();
+
+                return $this->redirectToRoute('user_index');
+            }
+
+            return $this->render('user/edit.html.twig', [
+                'user' => $user,
+                'form' => $form->createView(),
+            ]);
         }
-
-        return $this->render('user/edit.html.twig', [
-            'user' => $user,
-            'form' => $form->createView(),
-        ]);
+        catch(Exception $e){
+            return $this->redirectToRoute('home');
+        }
     }
 
     /**
@@ -91,12 +127,22 @@ class UserController extends AbstractController
      */
     public function delete(Request $request, User $user): Response
     {
-        if ($this->isCsrfTokenValid('delete'.$user->getId(), $request->request->get('_token'))) {
-            $entityManager = $this->getDoctrine()->getManager();
-            $entityManager->remove($user);
-            $entityManager->flush();
-        }
+        try {
+            if(!$this->getUser()) throw new AccessDeniedException();
+            $this->denyAccessUnlessGranted('ROLE_ADMIN');
 
-        return $this->redirectToRoute('user_index');
+            $form = $this->createForm(UserType::class, $user);
+            $form->handleRequest($request);
+            if ($this->isCsrfTokenValid('delete'.$user->getId(), $request->request->get('_token'))) {
+                $entityManager = $this->getDoctrine()->getManager();
+                $entityManager->remove($user);
+                $entityManager->flush();
+            }
+
+            return $this->redirectToRoute('user_index');
+        }
+        catch(Exception $e){
+            return $this->redirectToRoute('home');
+        }
     }
 }
